backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.18.4",
    "fastapi[standard]>=0.129.0",
    "pre-commit>=4.5.1",
    "psycopg[binary]>=3.3.3",
    "pwdlib[argon2]>=0.3.0",
    "pydantic>=2.12.5",
    "pyjwt>=2.11.0",
    "ruff>=0.15.2",
    "sqlmodel>=0.0.34",
    "ty>=0.0.17",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ty]
strict = true
warn-unused-ignores = true
no-implicit-optional = true
include = [
    "src/**"
]

[tool.ruff]
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
    "B",   # bugbear
    "C4",  # comprehensions
    "SIM", # simplifications
]
ignore = [
    "E501",
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
